Do not mutate weights in reduce.mode! Fixes #145
Apart from mutating the weights (which are eligilable for re-use!), this might also cause issues with concurrent writes, since the regrid uses a parallel for loop via numba.prange.

With this commit, we allocate a new (heap) array for every call of mode. For large grids, this is an enormous amount of small allocations, but mode is a relatively slow method anyway...